refactor: 공통 도메인 패키지명을 변경하라
