+++ src/main/java/kitchenpos/common/exception/ExceptionAdvice.java
@@ -1,4 +1,4 @@
-package kitchenpos.core.exception;
+package kitchenpos.common.exception;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
