+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -15,8 +15,8 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
-import kitchenpos.core.exception.CannotUpdateException;
-import kitchenpos.core.exception.ExceptionType;
+import kitchenpos.common.exception.CannotUpdateException;
+import kitchenpos.common.exception.ExceptionType;
 import kitchenpos.order.application.OrderValidator;
 import kitchenpos.order.domain.request.OrderRequest;
