+++ src/main/java/kitchenpos/table/application/TableService.java
@@ -1,9 +1,9 @@
 package kitchenpos.table.application;
 
 import java.util.stream.Collectors;
-import kitchenpos.core.exception.BadRequestException;
-import kitchenpos.core.exception.ExceptionType;
-import kitchenpos.core.exception.NotFoundException;
+import kitchenpos.common.exception.BadRequestException;
+import kitchenpos.common.exception.ExceptionType;
+import kitchenpos.common.exception.NotFoundException;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
 import kitchenpos.table.dto.request.OrderTableRequest;
