+++ src/main/java/kitchenpos/table/domain/TableUnGroupEventHandler.java
@@ -3,8 +3,8 @@ package kitchenpos.table.domain;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.core.exception.CannotUpdateException;
-import kitchenpos.core.exception.ExceptionType;
+import kitchenpos.common.exception.CannotUpdateException;
+import kitchenpos.common.exception.ExceptionType;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.tableGroup.domain.TableUnGroupEventPublisher;
