+++ src/main/java/kitchenpos/tableGroup/application/TableGroupService.java
@@ -1,7 +1,7 @@
 package kitchenpos.tableGroup.application;
 
-import kitchenpos.core.exception.ExceptionType;
-import kitchenpos.core.exception.NotFoundException;
+import kitchenpos.common.exception.ExceptionType;
+import kitchenpos.common.exception.NotFoundException;
 import kitchenpos.table.application.TableService;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.tableGroup.domain.TableGroupValidator;
