+++ src/test/java/kitchenpos/order/domain/OrderTest.java
@@ -6,11 +6,9 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.mock;
 
 import java.util.Collections;
-import kitchenpos.core.exception.CannotUpdateException;
-import kitchenpos.core.exception.ExceptionType;
+import kitchenpos.common.exception.CannotUpdateException;
+import kitchenpos.common.exception.ExceptionType;
 import kitchenpos.order.application.OrderValidator;
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.request.OrderRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
